Add Zizmor to checks workflow (#2)
* Add Zizmor to checks workflow

* Add extra permission for Zizmor step

* Don't persist credentials for checkout actions

* Make zizmor more pedantic

* Address new Zizmor alerts

Curiously, it looks like it only is finding problems in the job that runs
zizmor and not the rest of the workflow, which wouldn't be very helpful. I'll
see if it reports other issues after these changes.

* Specify permissions for other jobs in workflow

* Remove job name again, it seems a bit silly to require on one job

* Disabled advanced-security, enable annotations

This is to overcome the quite odd limitations of GitHub's Advanced Security
reporting, where issues aren't reported in case they happen on lines that
aren't fully inside the changed lines of a PR:
https://github.com/zizmorcore/zizmor-action#changes-introduce-security-alerts-but-no-pr-checks-are-shown.

* Specify concurrency settings

* Set min-severity to low, it means info level alerts are ignored